+++ site_scons/site_tools/separate_debug_symbols.py
@@ -1,59 +1,75 @@
 """
 Tool to separate debug symbols.
+
+Adds SeparateDebugSymbols pseudo-builder that registers a post-action for splitting
+symbols, with a cleanup.
 """
 
 from SCons.Script import Clean, Exit, Action
 import os
 
-
-def separate_debug_symbols(target, source, env):
+def separate_debug_symbols(target, source, env, p_symbols_path):
     platform = env["platform"]
 
-    target_path = str(target[0])
-    source_path = str(source[0])
-
-    def run(cmd):
-        err = os.system(cmd)
-        return os.WEXITSTATUS(err)
+    binary_path = str(target[0])
+    symbols_path = str(p_symbols_path)
 
     if platform in ["macos", "ios"]:
-        err = run(f'dsymutil "{source_path}" -o "{target_path}"')
+        err = env.Execute(f'dsymutil "{binary_path}" -o "{symbols_path}"')
         if err != 0:
             print(f"ERROR: Failed to split debug symbols (exit code {err})")
             Exit(1)
 
-        err = run(f'strip -u -r "{source_path}"')
+        err = env.Execute(f'strip -u -r "{binary_path}"')
         if err != 0:
             print(f"ERROR: Failed to strip debug symbols (exit code {err})")
             Exit(1)
+
     elif platform == "linux":
-        err = run(f'objcopy --only-keep-debug --compress-debug-sections=zlib "{source_path}" "{target_path}"')
+        err = env.Execute(f'objcopy --only-keep-debug --compress-debug-sections=zlib "{binary_path}" "{symbols_path}"')
         if err != 0:
             print(f"ERROR: Failed to split debug symbols (exit code {err})")
             Exit(1)
 
-        err = run(f'strip --strip-debug --strip-unneeded "{source_path}"')
+        err = env.Execute(f'strip --strip-debug --strip-unneeded "{binary_path}"')
         if err != 0:
             print(f"ERROR: Failed to strip debug symbols (exit code {err})")
             Exit(1)
 
-        err = run(f'objcopy --add-gnu-debuglink="{target_path}" "{source_path}"')
+        err = env.Execute(f'objcopy --add-gnu-debuglink="{symbols_path}" "{binary_path}"')
         if err != 0:
             print(f"ERROR: Failed to add debug link (exit code {err})")
             Exit(1)
+
+    elif platform == "android":
+        err = env.Execute(f'llvm-objcopy --only-keep-debug --compress-debug-sections=zlib "{binary_path}" "{symbols_path}"')
+        if err != 0:
+            print(f"ERROR: Failed to split debug symbols (exit code {err})")
+            Exit(1)
+
+        err = env.Execute(f'llvm-strip --strip-debug --strip-unneeded "{binary_path}"')
+        if err != 0:
+            print(f"ERROR: Failed to strip debug symbols (exit code {err})")
+            Exit(1)
+
+        err = env.Execute(f'llvm-objcopy --add-gnu-debuglink="{symbols_path}" "{binary_path}"')
+        if err != 0:
+            print(f"ERROR: Failed to add debug link (exit code {err})")
+            Exit(1)
+
     else:
         print("ERROR: Can't separate debug symbols on this platform")
         Exit(1)
 
 
-def command(env, target, source):
-    result = env.Command(
-        target,
-        source,
-        Action(separate_debug_symbols, cmdstr="Separating debug symbols: $SOURCE -> $TARGET")
-    )
-    Clean(target, target)
-    return result
+def command(env, symbols_path, binary):
+    # Closure captures symbols_path
+    def action(target, source, env):
+        separate_debug_symbols(target, source, env, symbols_path)
+
+    env.AddPostAction(binary, Action(action, cmdstr=f"Separating debug symbols: {binary} -> {symbols_path}"))
+    env.Clean(binary, symbols_path)
+    return binary
 
 
 def generate(env):